Move database helper functions to their own file
